refactor(hardware): types: polymorphic field ctors (#10401)
By making the class arg of BinaryFieldBase factory functions a typevar
with a bound of the base itself, when using the factories from child
classes of BinaryFieldBase (i.e. UInt8Field.build()) then the return
type will be the child class. This means, among other htings, that we
can use the base class factories in child class factories that are
explicitly noted to return instances of the child class.