feat(ot3-hardware): support tip action requests in the move group runner
Related to #RET-916 and #RET-918. Adds in TipActionRequest and TipActionResponse to the move group
runner as well as a script to test pick up and drop tip behavior on the 96 channel. For now, holding
off on adding this to the hardware controller until behavior is finalized around using the tip sense
detection.