refactor(api): ot3: settable cal thresholds (#10469)
* feat(hardware): allow setting the threshold offset

The sensors can now be set to an autobaseline + offset threshold mode,
rather than an autobaseline mode that has a hardcoded offset. By default
that threshold is 1 of whatever units the sensors use, which might or
might not actually be a reasonable value.

* refactor(api): ot3: sensor threshold offsets in config

These offsets can be used to tune the sensitivity of autocalibration.
They control the amount of extra capacitance the sensors must see over
their baseline to trigger.