refactor(hardware): capture sensor data
The sensor scheduler has a method to tell sensors to control the sync
line that can optionally log data, but sometimes we want to _capture_
the data and not touch the sync line at all. A new capture_output
context manager will capture streaming data from the pipette sensor and
yield it to an asyncio queue for later processing.