refactor(api): ot3: noncontact calibration
This calibration method, rather than doing a binary search to find the
edges of the slot, sweeps a single pass over the entirety of the slot in
an axis while recording the capacitive sensor's output. By moving at a
constant speed, we should get a trace that approximates the geometry of
the slot, with rolled-off edges corresponding to the area of the probe
face that remains above the deck. Because the probe face is circular,
the rate of change of the signal should be at its most extreme when the
center of the probe is over the slot edge, when the most area is passing
by the edge. We should therefore be able to find the extrema of the
finite difference of the capacitive sensor output and set the edges
equal to the extrema.

This commit is a prototype; more needs to be done to integrate this
properly, like moving a couple constants into config, and making carved
out elements of configuration for parts of the behavior rather than
reusing configuration elements used for the binary search protocol.