refactor(api,hardware): o3: add capacitive_pass (#10613)
* refactor(hardware): capture sensor data

The sensor scheduler has a method to tell sensors to control the sync
line that can optionally log data, but sometimes we want to _capture_
the data and not touch the sync line at all. A new capture_output
context manager will capture streaming data from the pipette sensor and
yield it to an asyncio queue for later processing.

* refactor(hardware): add capacitive_pass

This new method in tool_sensors works similarly to capacitive_probe, and
has similar arguments, but rather than stopping motion when a threshold
is exceeded it runs the entire motion and returns a list of all data
captured during the motion.

* refactor(api): ot3: add capacitive_sweep

This hardware control method sweeps the specified axis while capturing
capacitive sensor data.