refactor(api): ot3: noncontact calibration (#10617)
* refactor(hardware): split move prep and run

The MoveGroupRunner when executed provisions all moves to the axis
controllers and then runs them, all in the same coroutine. This is
usually fine, but sometimes callers want to do some other preparation
immediately before the execute message goes out. By allowing the prep
phase (sending all the move groups) and the execute phase (beginning the
motion) to be split, we can add this funcitonality.

* refactor(hardware): capture sensor data

The sensor scheduler has a method to tell sensors to control the sync
line that can optionally log data, but sometimes we want to _capture_
the data and not touch the sync line at all. A new capture_output
context manager will capture streaming data from the pipette sensor and
yield it to an asyncio queue for later processing.

* refactor(hardware): add capacitive_pass

This new method in tool_sensors works similarly to capacitive_probe, and
has similar arguments, but rather than stopping motion when a threshold
is exceeded it runs the entire motion and returns a list of all data
captured during the motion.

* refactor(api): ot3: add capacitive_sweep

This hardware control method sweeps the specified axis while capturing
capacitive sensor data.

* refactor(api): ot3: noncontact calibration

This calibration method, rather than doing a binary search to find the
edges of the slot, sweeps a single pass over the entirety of the slot in
an axis while recording the capacitive sensor's output. By moving at a
constant speed, we should get a trace that approximates the geometry of
the slot, with rolled-off edges corresponding to the area of the probe
face that remains above the deck. Because the probe face is circular,
the rate of change of the signal should be at its most extreme when the
center of the probe is over the slot edge, when the most area is passing
by the edge. We should therefore be able to find the extrema of the
finite difference of the capacitive sensor output and set the edges
equal to the extrema.

This commit is a prototype; more needs to be done to integrate this
properly, like moving a couple constants into config, and making carved
out elements of configuration for parts of the behavior rather than
reusing configuration elements used for the binary search protocol.

* refactor(api):

Log calibration results to json for later analysis

This probably needs some more changes to limit the number of files