feat(hardware): add ot3 gpio driver, set estop
Add a GPIO driver for the OT3. It's like the rpi_gpio driver, but OT3
drivers live in opentrons_hardware. There isn't a lot of duplicated
code.

The annoying thing is that libgpiod isn't really ever going to be
runnable in a test context, so the code automatically switches out for a
mock if libgpiod isn't available.

With libgpiod and with the effort toradex put in to giving gpios
appropriate semantic names, we can pull out the estop gpio line and set
it high, which allows motors to be enabled - on new hardware, with that
line connected, if you don't explicitly set that line high nothing moves.