refactor(api): motion_utilities uses typevars now
The reason that motion_utilities needed a bunch of overloads and
couldn't use typevar was that the OT-2 variant converted between string
axis names and axis enums in addition to converting between coordinate
systems, which can't be represented in a typevar. By moving the
name->axis conversion out of motion utilities, we can a) use typevars
and b) always rely on having an enum element as an input which means we
can quit passing the enum class in on its own.

While we're at it we can slightly adapt deck_from_machine to not
hardcode z axes which makes it adapt to the presence of a gripper Z axis.

more axis removal