refactor(types): better assignment of axis categories
feat(api): ot3: handle gripper mount in move

The functions that take mounts (and apply to the gripper - e.g., not
aspirate, pick_up_tip, etc) should handle the gripper mount and the
gripper Z axis. The positions of these axes should be in current
position and gantry position should let you specify the gripper to get
the position of. With the refactors to motion_utils, this isn't that
hard.

The problem is that this doesn't actually work with the protocols that
define the hardware api, because it's a different type of axis. So we
make some new functions that expose the new axes for now.

In the long term, those protocols need to be made generic on the axis
type and consumers need to use either an
OT3HardwareAPI(=HardwareControlAPI[OT3Axis, OT3Mount]) or
OT2HardwareAPI(=HardwareControlAPI[Axis, Mount]).