refactor(api): support gripper mount in hardware control (#10967)
* refactor(api): motion_utilities uses typevars now

The reason that motion_utilities needed a bunch of overloads and
couldn't use typevar was that the OT-2 variant converted between string
axis names and axis enums in addition to converting between coordinate
systems, which can't be represented in a typevar. By moving the
name->axis conversion out of motion utilities, we can a) use typevars
and b) always rely on having an enum element as an input which means we
can quit passing the enum class in on its own.

While we're at it we can slightly adapt deck_from_machine to not
hardcode z axes which makes it adapt to the presence of a gripper Z axis.

more axis removal

* refactor(types): better assignment of axis categories

feat(api): ot3: handle gripper mount in move

The functions that take mounts (and apply to the gripper - e.g., not
aspirate, pick_up_tip, etc) should handle the gripper mount and the
gripper Z axis. The positions of these axes should be in current
position and gantry position should let you specify the gripper to get
the position of. With the refactors to motion_utils, this isn't that
hard.

The problem is that this doesn't actually work with the protocols that
define the hardware api, because it's a different type of axis. So we
make some new functions that expose the new axes for now.

In the long term, those protocols need to be made generic on the axis
type and consumers need to use either an
OT3HardwareAPI(=HardwareControlAPI[OT3Axis, OT3Mount]) or
OT2HardwareAPI(=HardwareControlAPI[Axis, Mount]).

* add gripper axes to backend handlers

* actual correct gripper offsets

* ot3: add gripper node splitting and more tests

node split format

* second map for gripper model 1

model map format

* add CriticalPoint preimport to repl