fix(ot3): modify the device info integration test to check for all expected nodes during broadcast
The `run_simulators.sh` script used by the emulation nightly tests actually also runs a bootloader
with a pipette_left_bootloader NodeId. This will sometimes cause the `test_each_node` test to fail
in the `test_device_info.py` file because it is reading from the queued up can messages and does not
expect a pipette_left_bootloader message to be there.