chore(app): add python executable for arm64-linux
This is used in the builds for the OT-3 on-device display. It's kind of
gross to add another python interpreter to the robot build, and I think
in the future we should refactor the app to use the installed python
stuff in on-device mode, but with this option absent then `beforeBuild`
resolves to false when invoked from electron-builder, which silently
skips dependency gathering and results in a broken app, so this is a bit better.