chore(app-shell): build: optionally skip python
On builds that are creating apps for the robot, we need to not bundle
python - bundling python takes us down a long and winding road of trying
to cross-compile all the dependencies. Add a makefile argument that will
skip installing python, and change the before-build script so that it
raises an error if there is no python in the environment map if that arg
is not present.