fix(app): fix labware offset text alignment, fix historical run log missing timestamps
Ensure that labware names are left aligned in the LabwareInfoOverlay, even if an offset is present.
Ensure that historical runs which involved more than 60 LPC (or otherwise pre-play) run commands are
able to correctly glean the first post play run command in the absence of a current command key.
Remove unnecessary polling of the protocol record from the run detail page.

Closes #11109, Closes #11108