feat(hardware): convert encoder pulses to encoder position in mm (#10288)
Encoder pulses from axes that have them are now sent over in millimeters. They are signed values, to account for occasionally bouncing past the endpoint when homing.

The positions are available alongside the stepper estimated positions and there is now support for a GetPosition command that returns the current encoder estimates.