feat(update-server): advertise robotModel
The robot model is being advertised by the robot server's health
endpoint, which is good but not sufficient to robustly identify a
machine to a client across all its possible connection states and
styles.

This PR adds the robot model to
- The update server health endpoint
- a TXT record in the MDNS advertisement

This should allow identification of the machine model to be robust
across machines that are found on MDNS but unreachable; machines that
are connected with IPs with no MDNS advertisement; and everything in
between.

It does create a possible conflict (though the machine type is static
data that should not conflict). Resolution mechanisms in clients should
follow these rules:

- When more than one source (i.e. of update-server health, robot-server
health, and TXT record) is presenting a valid machine type, prioritize
in the order of 1) robot server; 2) update server; 3) MDNS
- If any source does not have the appropriate key (robot_model for
robot-server health, robotModel for
update-server health, the TXT record robotModel for MDNS), that source
may be assumed to be an OT-2
- If any source is not present, i.e. requests to /health or
/server/health fail or the MDNS advertisement is not found, it should be disregarded

Closes RCORE-66