feat(discovery-client): robust robot models
As of https://github.com/Opentrons/opentrons/pull/11295 , robots
advertise their robot model in
- a TXT record attached to their MDNS advertisements
- the health response presented by the update server
in addition to the presentation in the health response from the robot
server.

The discovery client now harvests (and presents proper types for) all
this data and sends it upstream to whatever is listening, with a new
advertisedModel field for the MDNS response, and a proper shape for the
health data from the update server.