refactor(hardware): cap sensor integration fix
The capacitive sensor on hardware has an offset handler that will
gradually grow if necessary. The simulator harness in firmware would
pretend the sensor was returning a value large enough to make the offset
change, so over time the sensor value returned on can would grow. Making
the value returned by the harness smaller fixes this issue, but means we
have to change the expected values locally.