Handle unknown pipette names and models
The hardware layer is responsible for making sure that values are inside
correct python types. It makes sure that even if a random int gets
returned for the pipette name, it gets mapped to unknown; and that a
model int gets mapped to a dotted-version string.

Once those exist, the hardware control needs to handle values that are
in the right type, but are semantically invalid. Here, the solution is
to catch the underlying python errors of e.g. "i dont know what this
model means so here's a keyerror" into a semantically relevant exceptions.