feat(hardware): Load pipette model from pipette serial (#11349)
* refactor(hardware): flash pipette serials

Adds a new module for dealing with pipette serial numbers and a tool for
flashing them and confirming they're correct.

The hardware layer is responsible for making sure that values are inside
correct python types. It makes sure that even if a random int gets
returned for the pipette name, it gets mapped to unknown; and that a
model int gets mapped to a dotted-version string.

Once those exist, the hardware control needs to handle values that are
in the right type, but are semantically invalid. Here, the solution is
to catch the underlying python errors of e.g. "i dont know what this
model means so here's a keyerror" into a semantically relevant exceptions.

Also,

* refactor(hardware): cap sensor integration fix

The capacitive sensor on hardware has an offset handler that will
gradually grow if necessary. The simulator harness in firmware would
pretend the sensor was returning a value large enough to make the offset
change, so over time the sensor value returned on can would grow. Making
the value returned by the harness smaller fixes this issue, but means we
have to change the expected values locally.

Closes RET-1168

Co-authored-by: Caila Marashaj <caila.marashaj@opentrons.com>