refactor(api): use deck def for slot positions
We have a nice deck definition system where we describe the physical
geometry of the deck in json. We used that for position in most
contexts, but when executing APIv2 protocols we used the old-style
direct calculation of slot offsets for the positions we load labware
into. This doesn't work for the OT-3.

By changing that logic to also depend on the deck definition, which is
then selected based on what machine is running, we can get correct
positioning for both machines.