feat(hardware): ot3: set per-axis max speeds (#11425)
Currently, we can specify a linear speed that an individual move cannot
exceed. This limit is generally given by higher level code and is more
about scientific concerns (avoiding shaking, for instance) than about
robotics concerns.

However, there are also speed limits that are based on robotics
concerns. Any given axis has a speed it may not exceed due to physical
limits - the speed at which its dynamic friction overcomes its available
torque, the speed at which it's no longer possible to represent that
speed correctly in binary, etc - and that limit, too, must be
represented.

These speeds are properly part of an axis's constraints, so we add them
there.

We can implement the speed limit by making the input speed provisioned
on each Move object the smaller of the linear speed limit and the
projected axis-speed-limit-based value when we make a new Move object at
the beginning of a blend pass.

This happens by decomposing the linear speed through the move's unit
vector and comparing each component to the individual axis limit. If the
limit is lower, we can apply ratio between the axis speed limit and the
axis linear speed component to scale down the given speed.

* refactor(api): ot3: axis max speeds in constraints

We already had per-axis max speed values in our configs, and we... just
weren't doing anything with them. Now, we pass them into the system
constraints used by the hardware motion planner.

* refactor(api): ot3: limit plunger axis max speed

The plunger axis' maximum representable speed is 47mm/s at its current
steps/mm. Limit it to a nice round number below that.

* fix(api): ot3: config updates

Update the axis speed limits now that they are being respected, and bump
the currents.

Closes RCORE-125