refactor(shared-data): no tests, fixtures on OT-2
First, the test directory in shared-data python bindings was getting
included in the sdist even with the package. In the api, we do a src
directory to prevent this. That would work here but it would be a lot,
and we're already overloading the sdist builder, so we can rmtree the
tests directory and get the same outcome.

Also, remove the fixture data when we're installing to the OT-2. We
don't want to do this for the normal case of sdists, since we might want it.