fix(api): do not trigger module attach on temporary device node paths (#11546)
inotify now gives us notifications for kernel-created "tempnode" variants of the serial /dev entries that we normally bind to modules. These are meant to exist to give userspace a chance to handle new devices before the kernel does and before other udev rules fire. Unfortunately, they're named by just appending .tmp-{TYPE}{MAJOR}:{MINOR} to whatever is about to be created, so when we try and create e.g. /dev/ot_module_magdeck0 udev will helpfully first create /dev/ot_module_magdeck0.tmp-c122:0, which inotify sees and tells us about, and which matches our module regex, so we try to create a module on it. Unfortunately, that device only exists for an extremely short period of time, and by the time we try and create the module (which we shouldn't be doing anyway) it's gone and we get an exception.

The fix for this is to add the capability for the module parser to recognize these tempnode paths and discard the extension. Since tempnodes get created just ahead of the actual symlink, we can rely on the matching symlink to be present.

Closes [RQA-296](https://opentrons.atlassian.net/browse/RQA-296)