change lerna settings and add release scripts
The release node script will generate changelogs using
conventional-changelog directly for a specific tag. It'll be used in a
github workflow that runs when tags are pushed, and create a github
release for the tag that will contain the changelogs. By uncoupling
changelogs from git repo content, we're closer to being able to source a
version from git.