Don't home gripper if engine is configured to use virtual gripper.
We don't expect this to make any functional difference. The engine should only use a virtual gripper during analysis--but then, it will also be using a simulating hardware controller, so homing the gripper would no-op, anyway.

However, we think this is conceptually more in-line with our vision for virtual pipettes, grippers, and modules. In an ideal world, a ProtocolEngine used for analysis wouldn't access a hardware API at all.

Co-Authored-By: Sanniti <sanniti@opentrons.com>