feat(build): Aspirationally update releasing.md
This commit updates RELEASING.md to the new process for running
releases, which are now based entirely on pushing tags - no more manual
changelog generation and trimming, no more bump prs since the version is
not checked in anywhere. This drastically reduces the sheer number of
steps, number of prs, and number of builds to wait for.