refactor(build): Move changelogs to github releases (#11804)
We're going to add a new release channel of the same robot stack as for the OT-2. That means it needs different tags, and that flow won't work at all with the way we currently do versioning. This PR is the first of a series that will move the robot stack's source of truth for versioning to git tag names and abandon lerna.

This PR removes the CHANGELOG.md files, which were based only on the robot stack unprefixed tags, and replaces them with a git workflow and js script that will

    Create a release for every tag push, annotated correctly for the project for which the tag defines a new version. The release will always be marked as a draft to give an opportunity for human review. Versions with prerelease metadata (i.e. -alpha, -beta, -candidate) will be marked as prerelease.
    Generate the conventional-changelog for that release and put it in the body of the release. The changelogs are generated relative to the last version of equal or higher production status (see below)

That means that we get to remove the step of "review the changelogs" from the bump PR part of the releasing workflow, and "edit and compress the changelogs" from the actual release workflow.
New release flow

We still need lerna (for now) but the changelogs no longer change when you run make bump/lerna version and therefore do not need to be committed in the bump PR. The workflow for version bumps now looks like:

    Make a bump branch: git checkout -b chore_bump-25.0.0-alpha.2
    Bump: make bump version=25.0.0-alpha.2
    Commit the changes, which should only be to package.json in all subprojects: git add **/package.json ; git commit -m 'chore(release): 25.0.0-alpha.2
    Push the PR
    Squash merge the PR
    Push the version tag: git tag v25.0.0-alpha.2 ; git push origin v25.0.0-alpha.2

How the new changelogs work

While more changes are still to come in the rest of this PR series, this PR removes the requirement to edit changelogs from the release flow. The changelogs are now automatically generated in a github release created when you push a tag for any project - that means protocol designer and labware library now automatically get github releases.

Changelogs are generated for a version based on the last version of the same project of equal or greater production status, where production > candidate > beta > alpha (candidate, beta, and alpha are semver prerelease metadata tags, while production is a version with no prerelease metadata).

Some examples:
The first alpha for 25.0.0

We are going to start a new release which adds a major feature where the robot can make you coffee. We create a release branch, create a bump branch, run make bump, and merge back the new version. We tag and push v25.0.0-alpha.0. The last robot software release was v24.9.2. That is a production release, which is more-production than alpha, so changelogs are generated between v25.0.0-alpha.0 and v24.9.2.
Another alpha for 25.0.0

QA finds some bugs, so we need another alpha. We fix the issues, do a bump, merge the bump, and push the tag v25.0.0-alpha.1. The last tag was v25.0.0-alpha.0, which is an equally-production tag, so the changelog is generated for the changes between v25.0.0-alpha.0.
A beta for 25.0.0

Our robot now brings you a cup of coffee on command, but we want to see if users have feedback (perhaps they prefer hazelnut? want their cream measured in fl.oz. instead of mL?) so we make a beta. Bump, merge, tag v25.0.0-beta.0. The last three tags in order are v25.0.0-alpha.1, v25.0.0-alpha.0, and v24.9.2. The two alpha tags are strictly less-production than beta, so changelogs are generated between v25.0.0-beta.0 and v24.9.2.
A surprise release candidate for PD 38.2.0

Protocol designer has learned how to specify coffee amounts in json protocols, so we want a release! It contains most of the same changes as the robot software, but not quite the same. We happen to make a release candidate at this point in the sequence. It's tagged protocol-designer@38.2.0-candidate-a. We ignore the v25.0.0-beta.0, etc tags because they're not protocol-designer@ tags; the last such tag is protocol-designer@37.0.0-candidate-c, so that's what we use.

(Sidebar: candidate-c is not a good prerelease tag because prerelease metadata should look like -stage.sequence not -stage-sequence. Strictly speaking, the prerelease stage here is candidate-c, not candidate c. Let's use candidate.c or even better candidate.3 going forward if we can).
Upshot

We don't have to manually collapse all the alpha changelogs for the actual release anymore.