chore(build): remove versions from js package json
The versions in package.json are now all 0.0.0-dev. Versions do not yet
come from git, but when they do they'll be injected into package bundles
by webpack. That means that the yarn-level peer dependencies to other
monorepo packages can't have explicit version requests in them - the
local packages at 0.0.0-dev can't fulfill 6.1.0 or whatever, and we
still have some old published packages on npm that will take priority.
So we specify symlink dependencies. This ends up with the same structure
on disk in node_modules as we used to have.

The new structure showed some problems in a couple different tests that
we fixed along the way.

This commit just removes the version from package.json - it doesn't
actually replace it with anything - so all js builds in this commit are
now 0.0.0-dev unconditionally.