chore(build): find js app versions from git
Javascript apps that injected their versions into their distribution
bundles based on package.json now do it from git. This is based on a
refactor and generalization of the version-from-git machinery that backs
up the changelogs.

Only some of our js apps are actually aware of their own version and
make the version accessible inside the app. These are app, app-shell,
discovery-client, labware-library, and protocol-designer. They all do it
by using a webpack DefinePlugin to either
override node.process.env to provide fake environment variables or
establish globals that exist within the webpack bundle. That happens in
webpack.config.js for those apps, so webpack.config.js now pulls the
version that it injects from git.

Since dev builds are webpacked, this is generally works there too. The
exception is the desktop app.

The desktop app uses electron-updater to do automatic updates.
electron-updater's understanding of the current version is based on the
electron App module's getVersion function, which pulls the version from
the package.json and can't be overridden for some reason.

In actual app builds (from make -C app-shell dist*), this is fine
because we can selectively override package.json fields with
configuration in electron-builder.config.js.

When running the app in dev mode, we just run `yarn electron`, which
will always try and pick up the version from app-shell package.json,
which is 0.0.0-dev. That means that when you do a dev build of the app,
the version in the settings tab will be correct (that's from webpack since
it's in our js) but the updater will think it's the old version and
you'll always get an update popup. We can live with that.

Sadly enough, the above all means that github workflows that build
versioned artifacts need to be cloned with fetch-depth 0 so they
actually get history and tags.