chore(app-shell): create an ot-3 app build
This is a new "build variant" (new terminology alert!) of the app
created when you build the app for the ot3 project. It gets a different
product name so you can install it side-by-side with the production
opentrons app, and a different app icon so you can tell it apart.