refactor(shared-data, api): make command schema from py models, reference in protocol schema v7
Create a script and make task `make command_schema version=X` that will generate new a command
schema from the pydantic models of the protocol engine. Protocol Schema v7 now references commands
via a fixed reference to this command schema. Add a snapshot test to the opentrons/api project to
ensure that the command schema has not changed with out an intentional re-running of the schema
generation task.