chore(build): Derive application versions from git (#11863)
* feat(build): Aspirationally update releasing.md

This commit updates RELEASING.md to the new process for running
releases, which are now based entirely on pushing tags - no more manual
changelog generation and trimming, no more bump prs since the version is
not checked in anywhere. This drastically reduces the sheer number of
steps, number of prs, and number of builds to wait for.

* fix(build): some changelogs crept back in

* refactor(build): do not use lerna for versions

Remove lerna as a top-level dep and put placeholder versions in all package.json

* refactor(build): find py versions from git

The versions for python packages are now injected from the latest git
tag detected in the repo history, rather than from an otherwise-useless
package.json. This is dependent on the project - for packages used in
multiple projects, you have to decide what project
you're building for. That's done right now through an environment
variable. Happily enough, there's only two major projects that build
python packages and they use different distribution formats and thus
different makefile recipes, and thus can have different defaults.

This also requires modifying a lot of the internals of how we do
versions and while I was at it I went back to using phony targets in
makefiles for wheels and sdists because yeah it means you have to
rebuild them but it also makes the makefile way way way faster to parse.

The now-unused package.jsons are removed, and since yarn now doesn't
need to know about the python projects and they don't have package.json
anymore they're pulled out of the workspace list.

Sadly enough, this change means that any workflow that needs to produce
versioned artifacts is going to need git history.

* chore(build): remove versions from js package json

The versions in package.json are now all 0.0.0-dev. Versions do not yet
come from git, but when they do they'll be injected into package bundles
by webpack. That means that the yarn-level peer dependencies to other
monorepo packages can't have explicit version requests in them - the
local packages at 0.0.0-dev can't fulfill 6.1.0 or whatever, and we
still have some old published packages on npm that will take priority.
So we specify symlink dependencies. This ends up with the same structure
on disk in node_modules as we used to have.

The new structure showed some problems in a couple different tests that
we fixed along the way.

This commit just removes the version from package.json - it doesn't
actually replace it with anything - so all js builds in this commit are
now 0.0.0-dev unconditionally.

* chore(build): find js app versions from git

Javascript apps that injected their versions into their distribution
bundles based on package.json now do it from git. This is based on a
refactor and generalization of the version-from-git machinery that backs
up the changelogs.

Only some of our js apps are actually aware of their own version and
make the version accessible inside the app. These are app, app-shell,
discovery-client, labware-library, and protocol-designer. They all do it
by using a webpack DefinePlugin to either
override node.process.env to provide fake environment variables or
establish globals that exist within the webpack bundle. That happens in
webpack.config.js for those apps, so webpack.config.js now pulls the
version that it injects from git.

Since dev builds are webpacked, this is generally works there too. The
exception is the desktop app.

The desktop app uses electron-updater to do automatic updates.
electron-updater's understanding of the current version is based on the
electron App module's getVersion function, which pulls the version from
the package.json and can't be overridden for some reason.

In actual app builds (from make -C app-shell dist*), this is fine
because we can selectively override package.json fields with
configuration in electron-builder.config.js.

When running the app in dev mode, we just run `yarn electron`, which
will always try and pick up the version from app-shell package.json,
which is 0.0.0-dev. That means that when you do a dev build of the app,
the version in the settings tab will be correct (that's from webpack since
it's in our js) but the updater will think it's the old version and
you'll always get an update popup. We can live with that.

Sadly enough, the above all means that github workflows that build
versioned artifacts need to be cloned with fetch-depth 0 so they
actually get history and tags.


Co-authored-by: Mike Cousins <mike@opentrons.com>